% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitocarta_functions.R
\name{assign_mitocarta_pathways}
\alias{assign_mitocarta_pathways}
\title{Assign MitoCarta Pathways}
\usage{
assign_mitocarta_pathways(gene_names = NULL, mitocarta_data = NULL)
}
\arguments{
\item{gene_names}{A vector of user-provided gene names which should match the
format provided for the \code{gene_id_type} provided in \code{format_mitocarta_data()}.
For gene symbols, \code{gene_names} should match the \code{mitocarta_id} column output
from \code{check_mitocarta_synonyms()}.}

\item{mitocarta_data}{A list with dataframe objects \code{MitoCarta_MitoGenes},
\code{MitoCarta_AllGenes}, and \code{MitoCarta_Pathways} generated from \code{extract_mitocarta_data()}}
}
\value{
A list of dataframes of the same format as \code{extract_mitocarta_data()}
with an an added dataframe \code{Assign_Pathways} which has the following columns:
\describe{
\item{Name}{Gene name, which will not be unique if it is assigned to multiple pathways}
\item{Pathway.Number}{If a gene is assigned to multiple pathways, this column
identifies the pathway number}
\item{Main.Path}{Full pathway description}
\item{SubPath1}{Primary pathway as derived from Main.Path}
\item{SubPath2}{Secondary pathway as derived from Main.Path, if it exists}
\item{SubPath3}{Tertiary pathway as derived from Main.Path, if it exists}
}
}
\description{
This function assigns and parses MitoCarta3.0 pathways to gene ids using the
output from  \code{extract_mitocarta_data()}, and if desired, \code{check_mitocarta_synonyms()}.
}
