% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitocarta_functions.R
\name{check_mitocarta_synonyms}
\alias{check_mitocarta_synonyms}
\title{Check MitoCarta Synonyms}
\usage{
check_mitocarta_synonyms(gene_names = NULL, mitocarta_data = NULL)
}
\arguments{
\item{gene_names}{A vector of user-provided gene names. The function will scan
the entire list of gene names to identify whether the gene name is listed as
either a primary gene or synonym as annotated by MitoCarta3.0}

\item{mitocarta_data}{A list with three dataframe objects \code{MitoCartaGenes},
\code{AllGenes}, and \code{MitoCartaPathways} generated from \code{extract_mitocarta_data()}
with or without formatting with \code{format_mitocarta_data()}
as long as \code{'gene_symbol'} was selected for \code{gene_id_type}.}
}
\value{
A dataframe of gene names, where column 1 (\code{given_id}) is the provided
\code{gene_names} and column 2 (\code{mitocarta_id}) is the is populated with the gene
names as annotated by MitoCarta3.0.
}
\description{
This function accepts a custom vector of \code{gene_symbols} and returns a parallel list of
genes with any necessary substitutions, should the gene name fall as a synonym
in the MitoCarta3.0 annotation. This is a cleaning step before pathway identification
to help match as many genes as possible.
}
