% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expected.counts.meta}
\alias{expected.counts.meta}
\title{Rat Gene Count Metadata}
\format{
\subsection{\code{expected.counts.meta}}{

A data frame with 32 observations of 7 variables
\describe{
\item{Core.ID}{Sample ID assigned by the University of Michigan Advanced Genomics
Core for linking sample data to \code{expected.counts}}
\item{Exp.ID}{Rat ID used for experimental identification}
\item{Line}{High Capacity Runner (HCR) or Low Capacity Runner (LCR) rat line identifier}
\item{Sex}{Female or male rat sex identifer}
\item{Treatment}{Water or saline, as control treatment identifiers from their
respective experiments}
\item{Name}{Matches Exp.ID}
\item{Exp}{Experiment number, which identifies the RNA-Sequencing batch}
}
}
}
\source{
University of Michigan
}
\usage{
expected.counts.meta
}
\description{
A dataframe containing the rat meta data used for DESeq experiment design
}
\keyword{datasets}
