% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expected.counts}
\alias{expected.counts}
\title{Rat Expected Gene Counts (RNA-Seq)}
\format{
\subsection{\code{expected.counts}}{

A data frame with 32883 observations of 32 variables where column names correspond
to \code{Core.ID} from \code{expected.counts.meta} and row names correspond to Ensembl
\code{gene_id} from \code{gene_name_conversion}
}
}
\source{
https://brcf.medicine.umich.edu/cores/advanced-genomics University of Michigan Advanced Genomics Core
}
\usage{
expected.counts
}
\description{
A dataframe containing the aligned counts from RNA-Sequencing
}
\keyword{datasets}
