% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_name_conversion}
\alias{gene_name_conversion}
\title{Gene Name Conversion}
\format{
\subsection{\code{gene_name_conversion}}{

A data frame with 32883 observations of 4 variables
\describe{
\item{gene_id}{Ensembl ID}
\item{entrez_id}{Entrez ID}
\item{external_gene_name}{NCBI Gene Name}
\item{description}{Gene description}
}
}
}
\source{
https://brcf.medicine.umich.edu/cores/advanced-genomics University of Michigan Advanced Genomics Core
}
\usage{
gene_name_conversion
}
\description{
A dataframe containing rattus norvegicus gene names as ENSEMBL IDs, Entrez IDs,
NCBI gene name, and gene descriptions
}
\keyword{datasets}
