% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitocarta_functions.R
\name{plot_mitocarta_pathways}
\alias{plot_mitocarta_pathways}
\title{Plot MitoCarta Pathways}
\usage{
plot_mitocarta_pathways(
  filtered_data = NULL,
  mitocarta_data = NULL,
  subpath = "Primary",
  fc_name = NULL
)
}
\arguments{
\item{filtered_data}{A dataframe with colnames \code{Name}, \code{SubPath1}, \code{SubPath2},
and \code{SubPath3} as generated from function \code{assign_mitocarta_pathways()} in the
\code{Assigned_Pathways} dataframe}

\item{mitocarta_data}{A list with dataframe objects \code{MitoCarta_MitoGenes},
\code{MitoCarta_AllGenes}, \code{MitoCarta_Pathways}, and \code{Assigned_Pathways} generated
from \code{extract_mitocarta_data()} and then \code{assign_mitocarta_data()}.}

\item{subpath}{One of \code{Primary} to visualize high-level pathways, or a primary
or secondary pathway name to indicate which MitoCarta pathway to parse
(i.e. \code{Metabolism} or \verb{Metals and cofactors}). Defaults to \code{Primary}. Is case
sensitive.}

\item{fc_name}{A character string indicating the column name in
\code{filtered_data} containing the logFC values to plot for the percent
pathway bar parts. If \code{NULL}, plot will not be populated with logFC colors.}
}
\value{
A list of ggplot image objects
}
\description{
This function makes bar graphs indicated the number of genes in various MitoCarta
pathways that are represented in a filtered dataset from the \code{assign_mitocarta_pathways()}
output.
}
